requirements.txt
aiofiles==0.8.0
anyio==3.6.2
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
discord-webhook==0.17.0
fastapi==0.72.0
flickrapi==2.4.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==6.2.5
python-dotenv==0.19.2
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.17.0
zipp==3.9.0
